backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langgraph
langchain-google-genai
python-dotenv
requests
